+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 2.1.0 - Unreleased
+### Added
+- Added experimental Bash completion support.
+  See instructions at the top of `./bash_completion/wren-deploy.sh`.
 
 ## 2.0.0 - 2018-03-10
 ### Added